check invalid pid when send signal to process group

+++ src/job.cpp
@@ -274,8 +274,7 @@ void JobObject::send(int sigNum) const {
     return;
   }
 
-  pid_t pid = this->getPid(0);
-  if (pid == getpgid(pid)) {
+  if (pid_t pid = this->getPid(0); pid > -1 && pid == getpgid(pid)) {
     kill(-pid, sigNum);
     return;
   }
